+++ CHANGELOG.md
@@ -21,6 +21,16 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
 
 ### Security
 
+## [1.0.0-rc.3](https://github.com/signalfx/splunk-otel-dotnet/releases/tag/v1.0.0-rc.3)
+
+This is a release candidate,
+built on top of [OpenTelemetry .NET Auto Instrumentation v1.0.0](https://github.com/open-telemetry/opentelemetry-dotnet-instrumentation/releases/tag/v1.0.0).
+
+### Changed
+
+- Updated [OpenTelemetry .NET Auto Instrumentation](https://github.com/open-telemetry/opentelemetry-dotnet-instrumentation):
+  [`1.0.0`](https://github.com/open-telemetry/opentelemetry-dotnet-instrumentation/releases/tag/v1.0.0).
+
 ## [1.0.0-rc.2](https://github.com/signalfx/splunk-otel-dotnet/releases/tag/v1.0.0-rc.2)
 
 This is a release candidate,